requirements.txt
python-camellia
ifstools
kbinxml